Fix research over-queueing crash and harden turn resolution
Queueing the same upgrade on consecutive turns was validated against the
lagging completed level only, so a player could research past the upgrade
table and crash the next fight with an IndexOutOfRangeException in
ExponentBonusFromUpgrades.

- Validate, GetNewResearches, and IsValid now count researches still in
  progress (and same-turn commands on other hexagons) toward the level,
  capped by both the hexagon offering and the defined upgrade table
- Queued researches use the next level for duration and cost
- NextTurn deducts the research cost from matter (was silently free)
- ResearchPanel disables research while one is in flight and shows it
- GameSession.ResolveTurn catches engine exceptions: the state and all
  staged orders are kept, the error is shown, and the debug menu becomes
  available for exporting the state instead of the circuit dying

+++ Source/Simulturn.Core/Model/State/StateValidation/ICommandValidation.cs
@@ -19,6 +19,11 @@ public record MissingArmyForMovement(string PlayerId, Hexagon Hexagon, Unit Unit
 
 public record MovementExceedsRange(string PlayerId, Hexagon Hexagon, Hexagon Destination, Unit Unit, int Distance, short Range) : ICommandValidation;
 
-public record UpgradeExceedsAvailableLevel(string PlayerId, Hexagon Hexagon, Upgrade Upgrade, byte CurrentLevel, int AvailableLevel) : ICommandValidation, IPlayerStateValidation;
+/// <summary>
+/// <paramref name="CurrentLevel"/> counts completed levels plus researches still in progress or
+/// commanded in the same turn; <paramref name="AvailableLevel"/> is the lower of the hexagon's
+/// researchable levels and the levels defined in <see cref="GameSettings.Upgrades"/>.
+/// </summary>
+public record UpgradeExceedsAvailableLevel(string PlayerId, Hexagon Hexagon, Upgrade Upgrade, int CurrentLevel, int AvailableLevel) : ICommandValidation, IPlayerStateValidation;
 
 public record UpgradeRequiresDot(string PlayerId, Hexagon Hexagon, Upgrade Upgrade) : ICommandValidation, IPlayerStateValidation;
